Added the State Extension and Components for displaying the contributors of a Repository
Added a new state layout:
There are now three entities. Repositories stores all repositories loaded from github.
Users stores all users loaded from github via the REST API.
Contributions are the relation between repositories and users. If a user contributed to a repository, then a contribution is added to the state.

The structure of the state is also represented in the angular component structure.
The RepositoryBrowserComponent displays a list of repositories. When clicking on a repository an effect is triggered that fetches the top contributors from github and adds them and the contribution relations to the store.
By toggling the repository, we open the RepositoryDetailComponent. Here currently only one detail is represented. For each contributing user the RepositoryContributorComponent is rendered, which just shows the login of the corresponding contributing user.

There is one efficiency problem. When opening fetching the contributions of a repository from the state, we filter through all contributions. We could make this more efficient in the feature, by providing another state entity which links repositories to a list of contribution ids. However this would make the code more complicated, so I emitted it.